Добавит в provider проверку наличия интернета и зареджектит промисы, если интернета нет.

+++ src/api/provider.js
@@ -1,29 +1,57 @@
+const isOnline = () => {
+  return window.navigator.onLine;
+};
+
 export default class Provider {
   constructor(api) {
     this._api = api;
   }
 
   getPoints() {
-    return this._api.getPoints();
+    if (isOnline()) {
+      return this._api.getPoints();
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   getDesinations() {
-    return this._api.getDesinations();
+    if (isOnline()) {
+      return this._api.getDesinations();
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   getOffers() {
-    return this._api.getOffers();
+    if (isOnline()) {
+      return this._api.getOffers();
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   createPoint(point) {
-    return this._api.createPoint(point);
+    if (isOnline()) {
+      return this._api.createPoint(point);
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   updatePoint(id, point) {
-    return this._api.updatePoint(id, point);
+    if (isOnline()) {
+      return this._api.updatePoint(id, point);
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   deletePoint(id) {
-    return this._api.deletePoint(id);
+    if (isOnline()) {
+      return this._api.deletePoint(id);
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 }
